// src/app/shared/interfaces/distributer-collection-form.ts
export interface DistributerCollectionForm {
    batchId: string;
    orderNumber: string;
    dispatchDate: Date;
    distributorId: string;
    retailerId: string;
    fleetId: string;
    vehicleNumber: string;
    fromLocation: string;
    toLocation: string;
    journeyStartDate: Date;
    driverName: string;
    driverContactNumber: string;
    note: string;
}